+++ transfer.go
@@ -79,7 +79,7 @@ func NewTransfer(s *Session, atp AddTransferParams) (*Transfer, error) {
 		size:               atp.Size,
 		numPieces:          int(DivCeil(atp.Size, PieceSize)),
 		filePath:           atp.FilePath,
-		finalName:          strings.TrimSpace(atp.FinalName),
+		finalName:          sanitizeTransferFinalName(atp.FinalName),
 		fileComment:        strings.TrimSpace(atp.FileComment),
 		stat:               NewStatistics(),
 		closedStat:         NewStatistics(),
@@ -129,6 +129,12 @@ func NewTransfer(s *Session, atp AddTransferParams) (*Transfer, error) {
 	} else {
 		t.state = Downloading
 		t.markResumeDirty()
+	}
+	if t.IsFinished() && s != nil && s.settings.UseEmuleTempLayout && isEmuleTempPartPath(t.GetFilePath()) {
+		if t.handler != nil {
+			_ = t.handler.Close()
+		}
+		t.promoteEmuleTempPartIfNeeded()
 	}
 
 	return t, nil
@@ -902,7 +908,15 @@ func (t *Transfer) OnBlockRestoreCompleted(block data.PieceBlock, ec BaseErrorCo
 	}
 }
 
-func (t *Transfer) OnReleaseFile(_ BaseErrorCode, _ [][]byte, _ bool) {}
+func (t *Transfer) OnReleaseFile(_ BaseErrorCode, _ [][]byte, deleteFile bool) {
+	if deleteFile || !t.isFinishedForSharePublish() {
+		return
+	}
+	t.promoteEmuleTempPartIfNeeded()
+	if t.session != nil {
+		t.session.tryAddCompletedTransferToSharedStore(t)
+	}
+}
 
 func (t *Transfer) finished() {
 	for _, c := range t.connections {
@@ -912,13 +926,8 @@ func (t *Transfer) finished() {
 	}
 	t.connections = nil
 	t.state = Finished
-	if t.handler != nil {
-		_ = t.handler.Close()
-	}
-	t.promoteEmuleTempPartIfNeeded()
 	if t.session != nil {
 		t.session.SubmitDiskTask(NewAsyncRelease(t, false))
-		t.session.tryAddCompletedTransferToSharedStore(t)
 		t.session.PublishTransferToServer(t)
 		t.session.PublishTransferToKAD(t)
 		t.session.PublishTransferToKADV6(t)
@@ -942,13 +951,26 @@ func (t *Transfer) promoteEmuleTempPartIfNeeded() {
 	if incoming := strings.TrimSpace(t.session.settings.IncomingDir); incoming != "" {
 		destDir = incoming
 	}
-	dest, err := promoteEmulePartFile(src, destDir, name, t.size)
+	dest, err := promoteEmulePartFile(src, destDir, name)
 	if err != nil {
 		debugPeerf("transfer %s promote part failed: %v", t.hash.String(), err)
 		return
 	}
+	if dest == src {
+		return
+	}
 	t.filePath = dest
-	t.handler = disk.NewDesktopFileHandler(dest)
+	if t.handler != nil {
+		_ = t.handler.Close()
+	}
+	if setter, ok := t.handler.(interface{ SetPath(string) error }); ok {
+		if err := setter.SetPath(dest); err != nil {
+			debugPeerf("transfer %s set handler path: %v", t.hash.String(), err)
+		}
+	}
+	if t.session.sharedStore != nil {
+		t.session.sharedStore.UpdatePath(t.hash, dest, t.FileName())
+	}
 }
 
 func (t *Transfer) AsyncRestoreBlock(block data.PieceBlock) {